Implement the endpoint methods

+++ springsecurity/src/main/java/com/samsonmarikwa/restws/service/ProductService.java
@@ -17,6 +17,6 @@ public interface ProductService {
 	
 	@Path("/products")
 	@POST
-	int addProduct(Product product);
+	long addProduct(Product product);
 	
 }
